% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calcDefCapital.R
\name{calcDefCapital}
\alias{calcDefCapital}
\title{Calculates the Default Capital Charge}
\usage{
calcDefCapital(trades, EAD, reg_data, effective_maturity)
}
\arguments{
\item{trades}{The full list of the Trade Objects}

\item{EAD}{The Exposure-At-Default of the trades as per the selected regulatory framework}

\item{reg_data}{A list containing data related to the regulatory calculations (for example the regulatory probability-of-default, the regulatory loss-given-default etc)}

\item{effective_maturity}{The effective maturity of the trades of the netting set}
}
\value{
The default capital charge
}
\description{
Calculates the default capital charge using the advanced IRB methodology and the stressed R
}
\author{
Tasos Grivas <tasos@openriskcalculator.com>
}

