% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcNGR.R
\name{CalcNGR}
\alias{CalcNGR}
\title{Calculates the Net/Gross ratio (NGR)}
\usage{
CalcNGR(MtM_Vector)
}
\arguments{
\item{MtM_Vector}{A vector containing the trades to be netted}
}
\value{
The Net-Gross ratio (NGR)
}
\description{
Calculates the Net/Gross ratio used under the CEM regulatory framework
}
\author{
Tasos Grivas <tasos@openriskcalculator.com>
}
