% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcPD.R
\name{CalcPD}
\alias{CalcPD}
\title{Calculates the Probablity of Default}
\usage{
CalcPD(spread, LGD, time_points)
}
\arguments{
\item{spread}{The spread based on the credit curve}

\item{LGD}{The loss-given-default}

\item{time_points}{The timepoints that the analysis is performed on}
}
\value{
A vector containing the probablity of default on the specified timepoints
}
\description{
Calculates the probablity of the default on specific time points by using the spread of the corresponding credit curve and the loss given default
}
\author{
Tasos Grivas <tasos@openriskcalculator.com>
}
