% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcSimulatedExposure.R
\name{CalcSimulatedExposure}
\alias{CalcSimulatedExposure}
\title{Calculated the Simulated Exposure Profile}
\usage{
CalcSimulatedExposure(
  discount_factors,
  time_points,
  spot_curve,
  CSA,
  trades,
  sim_data
)
}
\arguments{
\item{discount_factors}{The discount curve derived from the spot curve}

\item{time_points}{The timepoints that the analysis is performed on}

\item{spot_curve}{The curve derived from interpolating the market spot rates}

\item{CSA}{The margin agreement}

\item{trades}{The list of the trade objects}

\item{sim_data}{A list containing simulation-related data (model parameters and number of simulation)}
}
\value{
A list containing the exposure profile (both collateralized and uncollateralized)
}
\description{
Calculates the simulated exposure profile (EE, NEE, PFE, EEE) by use of the Hull-White model. Two sets of results are provided:
one after taking into account the marging agreement and one assuming that there is no marging agreement present
}
\author{
Tasos Grivas <tasos@openriskcalculator.com>
}
