% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalcVA.R
\name{CalcVA}
\alias{CalcVA}
\title{Calculates the Valuation Adjustment}
\usage{
CalcVA(exposure, discount_factors, PD, LGD)
}
\arguments{
\item{exposure}{A vector containing the exposure values on which the credit risk adjustment will be calculated}

\item{discount_factors}{The Discount Curve}

\item{PD}{The probability-of-Default}

\item{LGD}{The Loss-Given-Default}
}
\value{
The Valuation Adjustment Value
}
\description{
Calculates the Valuation Adjustment based on the exposure, the probability-of-default and the loss-given-default
}
\author{
Tasos Grivas <tasos@openriskcalculator.com>
}

