% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Curve.R
\docType{class}
\name{Curve-class}
\alias{Curve}
\title{Curve Class}
\arguments{
\item{Tenors}{The Tenors of the curve}

\item{Rates}{The rates on the corresponding tenors}

\item{interp_function}{(Optional) The interpolation function of the curve. Can be populated via the 'CalcInterpPoints' method}
}
\value{
An object of type Curve
}
\description{
Creates a Curve Object containing pairs of Tenors with relevant rates and the interpolation function.
Also, methods for populating the object via a .csv file and the generation of the interpolation function via cubic splines are included.
}
\author{
Tasos Grivas <tasos@openriskcalculator.com>
}

