% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/HashTable.R
\docType{class}
\name{HashTable-class}
\alias{HashTable}
\title{Hashtable Class}
\arguments{
\item{keys}{A vector of keys}

\item{values}{A vector of values mapping to the keys}

\item{keys_type}{The type of the keys}
}
\value{
values_type  The type of the values
}
\description{
Creates a hashtable-like object so as to represent data with a key structure (for example addon tables, rating-based factors etc).
Also, it includes methods for populating the object via a .csv file and finding a value based on a specific key on an interval of keys
}
\author{
Tasos Grivas <tasos@openriskcalculator.com>
}

