% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IS_ELIGIBLE_CCY.R
\name{IS_ELIGIBLE_CCY}
\alias{IS_ELIGIBLE_CCY}
\title{Checks if specified currency is low risk}
\usage{
IS_ELIGIBLE_CCY(ccy)
}
\arguments{
\item{ccy}{The currency to be checked}
}
\value{
TRUE if the currency is is eligible to receive reduced regulatory risk weights
}
\description{
Checks if the specified currency is eligible to receive reduced regulatory risk weights
}
\examples{

TRUE == IS_ELIGIBLE_CCY('EUR')

}
\references{
https://www.bis.org/basel_framework/chapter/MAR/50.htm?inforce=20230101&published=20200708
}
\author{
Tasos Grivas <tasos@openriskcalculator.com>
}
