% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadSupervisoryCVAData.R
\name{LoadSupervisoryCVAData}
\alias{LoadSupervisoryCVAData}
\title{Supervisory Data Loading}
\usage{
LoadSupervisoryCVAData()
}
\value{
A list with the required data
}
\description{
Loads the supervisory data (factors, correlation and option volatility)
for each Asset Class and SubClass
}
\references{
MAR50 - Credit Value Adjustment Framework
https://www.bis.org/basel_framework/chapter/MAR/50.htm?inforce=20230101&published=20200708
}
\author{
Tasos Grivas <tasos@openriskcalculator.com>
}
