% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/xVACalculatorExample.R
\name{xVACalculatorExample}
\alias{xVACalculatorExample}
\title{xVA calculation example}
\usage{
xVACalculatorExample()
}
\value{
A list with the values of various valuations' adjustments
}
\description{
Calculates the xVA values for a simple example containing two IR swaps.
}
\author{
Tasos Grivas <tasos@openriskcalculator.com>
}

